Added Windows C demo
